WIP test refactors
