+++ solana-programs/reward-manager/program/tests/change_reward_manager_authority.rs
@@ -2,7 +2,7 @@
 
 mod utils;
 use audius_reward_manager::instruction;
-use solana_sdk::signature::Keypair;
+use solana_sdk::{signature::Keypair, transaction::TransactionError, transport::TransportError};
 use utils::program_test;
 
 use solana_program::program_pack::Pack;
@@ -11,38 +11,17 @@ use solana_sdk::{signature::Signer, transaction::Transaction};
 use utils::*;
 
 #[tokio::test]
-async fn success_change_manager_authority() {
-    let program_test = program_test();
-
-    let mint = Keypair::new();
-    let mint_authority = Keypair::new();
-    let token_account = Keypair::new();
-
-    let reward_manager = Keypair::new();
-    let manager_account = Keypair::new();
-    let min_votes = 3;
-
-    let mut context = program_test.start_with_context().await;
-    let rent = context.banks_client.get_rent().await.unwrap();
-
-    create_mint(
-        &mut context,
-        &mint,
-        rent.minimum_balance(spl_token::state::Mint::LEN),
-        &mint_authority.pubkey(),
-    )
-    .await
-    .unwrap();
-
-    init_reward_manager(
-        &mut context,
-        &reward_manager,
-        &token_account,
-        &mint.pubkey(),
-        &manager_account.pubkey(),
+/// Registered manager account can change the rewards manager manager
+async fn success_change_manager_() {
+    let TestConstants { 
+        reward_manager,
+        mut context,
+        token_account,
+        manager_account,
         min_votes,
-    )
-    .await;
+        ..
+    } = setup_test_environment().await;
+
 
     let new_manager = Keypair::new();
 
@@ -76,38 +55,13 @@ async fn success_change_manager_authority() {
 }
 
 #[tokio::test]
-async fn failure_change_manager_authority_bad_manager() {
-    let program_test = program_test();
-
-    let mint = Keypair::new();
-    let mint_authority = Keypair::new();
-    let token_account = Keypair::new();
-
-    let reward_manager = Keypair::new();
-    let manager_account = Keypair::new();
-    let min_votes = 3;
-
-    let mut context = program_test.start_with_context().await;
-    let rent = context.banks_client.get_rent().await.unwrap();
-
-    create_mint(
-        &mut context,
-        &mint,
-        rent.minimum_balance(spl_token::state::Mint::LEN),
-        &mint_authority.pubkey(),
-    )
-    .await
-    .unwrap();
-
-    init_reward_manager(
-        &mut context,
-        &reward_manager,
-        &token_account,
-        &mint.pubkey(),
-        &manager_account.pubkey(),
-        min_votes,
-    )
-    .await;
+/// Tries to change a manager, but passing in the incorrect reward_manager
+async fn failure_change_manager_bad_manager() {
+    let TestConstants { 
+        mut context,
+        manager_account,
+        ..
+    } = setup_test_environment().await;
 
     let new_manager = Keypair::new();
 
@@ -128,46 +82,23 @@ async fn failure_change_manager_authority_bad_manager() {
 
     let tx_result = context.banks_client.process_transaction(tx).await;
     match tx_result {
-        Err(e) if e.to_string() == "transport transaction error: Error processing Instruction 0: invalid account data for instruction" => return (),
-        Err(_) => panic!("Returned incorrect error!"),
-        Ok(_) => panic!("Incorrectly returned Ok!"),
+        Err(TransportError::TransactionError(TransactionError::InstructionError(0, solana_program::instruction::InstructionError::InvalidAccountData))) => assert!(true),
+        _ => panic!("Returned bad error!")
     }
 }
 
 #[tokio::test]
 #[should_panic(expected = "Transaction::sign failed with error KeypairPubkeyMismatch")]
+
+/// Tries to change a manager, but passes in a current manager which isn't 
+/// registered as manager
 async fn failure_change_manager_authority_bad_authority() {
-    let program_test = program_test();
-
-    let mint = Keypair::new();
-    let mint_authority = Keypair::new();
-    let token_account = Keypair::new();
-
-    let reward_manager = Keypair::new();
-    let manager_account = Keypair::new();
-    let min_votes = 3;
-
-    let mut context = program_test.start_with_context().await;
-    let rent = context.banks_client.get_rent().await.unwrap();
-
-    create_mint(
-        &mut context,
-        &mint,
-        rent.minimum_balance(spl_token::state::Mint::LEN),
-        &mint_authority.pubkey(),
-    )
-    .await
-    .unwrap();
-
-    init_reward_manager(
-        &mut context,
-        &reward_manager,
-        &token_account,
-        &mint.pubkey(),
-        &manager_account.pubkey(),
-        min_votes,
-    )
-    .await;
+    let TestConstants { 
+        reward_manager,
+        mut context,
+        manager_account,
+        ..
+    } = setup_test_environment().await;
 
     let new_manager = Keypair::new();
 
@@ -186,5 +117,5 @@ async fn failure_change_manager_authority_bad_authority() {
         context.last_blockhash,
     );
 
-    context.banks_client.process_transaction(tx).await;
+    context.banks_client.process_transaction(tx).await.unwrap();
 }
