+++ solana-programs/reward-manager/program/tests/transfer.rs
@@ -3,142 +3,39 @@ mod utils;
 
 use audius_reward_manager::{error::AudiusProgramError, instruction, processor::{SENDER_SEED_PREFIX, TRANSFER_ACC_SPACE, TRANSFER_SEED_PREFIX, VERIFY_TRANSFER_SEED_PREFIX}, state::{VerifiedMessages}, utils::{find_derived_pair, EthereumAddress}, vote_message};
 use libsecp256k1::{PublicKey, SecretKey};
-use rand::{Rng, prelude::ThreadRng, thread_rng};
-use solana_program::{instruction::{Instruction}, program_pack::Pack, pubkey::Pubkey, rent::Rent};
+use solana_program::{instruction::{Instruction}, program_pack::Pack, pubkey::Pubkey};
 use solana_program_test::*;
 use solana_sdk::{secp256k1_instruction::*, signature::Keypair, signer::Signer, transaction::{Transaction}};
 use std::{mem::MaybeUninit};
+use rand::{Rng};
 use utils::*;
 
-async fn setup_test_environment<'a>() -> (Keypair, [u8; 128], SecretKey, [u8; 128], ProgramTestContext, &'a str, Pubkey, Keypair, [u8;20], Keypair, Rent, ThreadRng, Keypair) {
-    /* Create verified messages and initialize reward manager */
-    let mut program_test = program_test();
-
-    program_test.add_program("claimable_tokens", claimable_tokens::id(), None);
-    let mut rng = thread_rng();
-
-    let mut context = program_test.start_with_context().await;
-
-    let mint = Keypair::new();
-    let mint_authority = Keypair::new();
-
-    let token_account = Keypair::new();
-    let reward_manager = Keypair::new();
-    let manager_account = Keypair::new();
-
-    let rent = context.banks_client.get_rent().await.unwrap();
-
-    create_mint(
-        &mut context,
-        &mint,
-        rent.minimum_balance(spl_token::state::Mint::LEN),
-        &mint_authority.pubkey(),
-    )
-    .await
-    .unwrap();
-
-    init_reward_manager(
-        &mut context,
-        &reward_manager,
-        &token_account,
-        &mint.pubkey(),
-        &manager_account.pubkey(),
-        3,
-    )
-    .await;
-
-    // Generate data and create oracle
-    let key: [u8; 32] = rng.gen();
-    let oracle_priv_key = SecretKey::parse(&key).unwrap();
-    let secp_oracle_pubkey = PublicKey::from_secret_key(&oracle_priv_key);
-    let eth_oracle_address = construct_eth_pubkey(&secp_oracle_pubkey);
-    let oracle_operator: EthereumAddress = rng.gen();
-
-    let oracle_derived_address = get_oracle_address(&reward_manager, eth_oracle_address);
-    
-    create_sender(
-        &mut context,
-        &reward_manager.pubkey(),
-        &manager_account,
-        eth_oracle_address,
-        oracle_operator,
-    )
-    .await;
-
-    let tokens_amount = 10_000u64;
-    let recipient_eth_key = [7u8; 20];
-    let transfer_id = "4r4t23df32543f55";
-
-    mint_tokens_to(
-        &mut context,
-        &mint.pubkey(),
-        &token_account.pubkey(),
-        &mint_authority,
-        tokens_amount,
-    )
-    .await
-    .unwrap();
-
-    let bot_oracle_message = vote_message!([
-        recipient_eth_key.as_ref(),
-        b"_",
-        tokens_amount.to_le_bytes().as_ref(),
-        b"_",
-        transfer_id.as_ref(),
-    ]
-    .concat());
-
-    let senders_message = vote_message!([
-        recipient_eth_key.as_ref(),
-        b"_",
-        tokens_amount.to_le_bytes().as_ref(),
-        b"_",
-        transfer_id.as_ref(),
-        b"_",
-        eth_oracle_address.as_ref(),
-    ]
-    .concat());
-
-    return (
-        reward_manager,
-        bot_oracle_message,
-        oracle_priv_key,
-        senders_message,
-        context,
-        transfer_id,
-        oracle_derived_address,
-        mint, 
-        recipient_eth_key,
-        token_account,
-        rent,
-        rng,
-        manager_account
-    )
-}
 
 #[tokio::test]
 async fn success_transfer() {
-    let (reward_manager,
+    let TestConstants { 
+        reward_manager,
         bot_oracle_message,
         oracle_priv_key,
         senders_message,
         mut context,
         transfer_id,
         oracle_derived_address,
-        mint,
         recipient_eth_key,
         token_account,
-        rent, 
+        rent,
         mut rng,
-        manager_account
-    ) = setup_test_environment().await;
+        manager_account,
+        recipient_sol_key,
+        ..
+    } = setup_test_environment().await;
 
     // Generate data and create senders
     let keys: [[u8; 32]; 3] = rng.gen();
     let operators: [EthereumAddress; 3] = rng.gen();
     let mut signers: [Pubkey; 3] = unsafe { MaybeUninit::zeroed().assume_init() };
     for (i, key) in keys.iter().enumerate() {
-        let derived_address = create_sender_from(&reward_manager, &manager_account, &mut context,key, operators[i]).await;
+        let derived_address = create_sender_from(&reward_manager, &manager_account, &mut context, key, operators[i]).await;
         signers[i] = derived_address;
     }
 
@@ -198,16 +95,6 @@ async fn success_transfer() {
     let transfer_account = get_transfer_account(&reward_manager, transfer_id);
     let verified_messages_account = get_messages_account(&reward_manager, transfer_id);
 
-    let recipient_sol_key = claimable_tokens::utils::program::get_address_pair(
-        &claimable_tokens::id(),
-        &mint.pubkey(),
-        recipient_eth_key,
-    )
-    .unwrap();
-    println!("Creating...Recipient sol key = {:?}", &recipient_sol_key.derive.address);
-    create_recipient_with_claimable_program(&mut context, &mint.pubkey(), recipient_eth_key).await;
-    println!("Created recipient sol key = {:?}", &recipient_sol_key.derive.address);
-
     let tx = Transaction::new_signed_with_payer(
         &[instruction::evaluate_attestations(
             &audius_reward_manager::id(),
@@ -249,7 +136,8 @@ async fn success_transfer() {
 /// wipe the account by calling `submit` again with correct attestations and
 /// finally succeed in `evaluate`.
 async fn invalid_messages_are_wiped() {
-    let (reward_manager,
+    let TestConstants {
+        reward_manager,
         bot_oracle_message,
         oracle_priv_key,
         senders_message,
@@ -259,10 +147,10 @@ async fn invalid_messages_are_wiped() {
         mint,
         recipient_eth_key,
         token_account,
-        _, 
         mut rng,
-        manager_account
-    ) = setup_test_environment().await;
+        manager_account,
+        ..
+    } = setup_test_environment().await;
 
     // Generate data and create senders
     let keys: [[u8; 32]; 4] = rng.gen();
@@ -433,89 +321,24 @@ async fn invalid_messages_are_wiped() {
 
 
 #[tokio::test]
+/// Purposefully send a corrupted DN attestation
 async fn failure_transfer_invalid_message_format() {
-    /* Create verified messages and initialize reward manager */
-    let mut program_test = program_test();
-
-    program_test.add_program("claimable_tokens", claimable_tokens::id(), None);
-    let mut rng = thread_rng();
-
-    let mut context = program_test.start_with_context().await;
-
-    let mint = Keypair::new();
-    let mint_authority = Keypair::new();
-
-    let token_account = Keypair::new();
-    let reward_manager = Keypair::new();
-    let manager_account = Keypair::new();
-
-    let rent = context.banks_client.get_rent().await.unwrap();
-
-    create_mint(
-        &mut context,
-        &mint,
-        rent.minimum_balance(spl_token::state::Mint::LEN),
-        &mint_authority.pubkey(),
-    )
-    .await
-    .unwrap();
-
-    init_reward_manager(
-        &mut context,
-        &reward_manager,
-        &token_account,
-        &mint.pubkey(),
-        &manager_account.pubkey(),
-        3,
-    )
-    .await;
-
-    // Generate data and create oracle
-    let key: [u8; 32] = rng.gen();
-    let oracle_priv_key = SecretKey::parse(&key).unwrap();
-    let secp_oracle_pubkey = PublicKey::from_secret_key(&oracle_priv_key);
-    let eth_oracle_address = construct_eth_pubkey(&secp_oracle_pubkey);
-    let oracle_operator: EthereumAddress = rng.gen();
-
-    let (_, oracle_derived_address, _) = find_derived_pair(
-        &audius_reward_manager::id(),
-        &reward_manager.pubkey(),
-        [SENDER_SEED_PREFIX.as_ref(), eth_oracle_address.as_ref()]
-            .concat()
-            .as_ref(),
-    );
-
-    create_sender(
-        &mut context,
-        &reward_manager.pubkey(),
-        &manager_account,
-        eth_oracle_address,
-        oracle_operator,
-    )
-    .await;
-
-    let tokens_amount = 10_000u64;
-    let recipient_eth_key = [7u8; 20];
-    let transfer_id = "4r4t23df32543f55";
-
-    mint_tokens_to(
-        &mut context,
-        &mint.pubkey(),
-        &token_account.pubkey(),
-        &mint_authority,
+    let TestConstants {
+        reward_manager,
+        bot_oracle_message,
+        oracle_priv_key,
+        mut context,
+        transfer_id,
+        oracle_derived_address,
+        mint,
+        recipient_eth_key,
+        token_account,
+        mut rng,
+        manager_account,
         tokens_amount,
-    )
-    .await
-    .unwrap();
-
-    let bot_oracle_message = vote_message!([
-        recipient_eth_key.as_ref(),
-        b"_",
-        tokens_amount.to_le_bytes().as_ref(),
-        b"_",
-        transfer_id.as_ref(),
-    ]
-    .concat());
+        eth_oracle_address,
+        ..
+     } = setup_test_environment().await;
 
     // Use invalid message format
     let senders_message = vote_message!([
@@ -533,34 +356,9 @@ async fn failure_transfer_invalid_message_format() {
     let keys: [[u8; 32]; 3] = rng.gen();
     let operators: [EthereumAddress; 3] = rng.gen();
     let mut signers: [Pubkey; 3] = unsafe { MaybeUninit::zeroed().assume_init() };
-    for item in keys.iter().enumerate() {
-        let sender_priv_key = SecretKey::parse(item.1).unwrap();
-        let secp_pubkey = PublicKey::from_secret_key(&sender_priv_key);
-        let eth_address = construct_eth_pubkey(&secp_pubkey);
-
-        let (_, derived_address, _) = find_derived_pair(
-            &audius_reward_manager::id(),
-            &reward_manager.pubkey(),
-            [SENDER_SEED_PREFIX.as_ref(), eth_address.as_ref()]
-                .concat()
-                .as_ref(),
-        );
-
-        signers[item.0] = derived_address;
-    }
-
-    for item in keys.iter().enumerate() {
-        let sender_priv_key = SecretKey::parse(item.1).unwrap();
-        let secp_pubkey = PublicKey::from_secret_key(&sender_priv_key);
-        let eth_address = construct_eth_pubkey(&secp_pubkey);
-        create_sender(
-            &mut context,
-            &reward_manager.pubkey(),
-            &manager_account,
-            eth_address,
-            operators[item.0],
-        )
-        .await;
+    for (i, key) in keys.iter().enumerate() {
+        let derived_address = create_sender_from(&reward_manager, &manager_account, &mut context, key, operators[i]).await;
+        signers[i] = derived_address;
     }
 
     let mut instructions = Vec::<Instruction>::new();
@@ -615,16 +413,7 @@ async fn failure_transfer_invalid_message_format() {
 
     context.banks_client.process_transaction(tx).await.unwrap();
 
-    let (_, verified_messages_derived_address, _) = find_derived_pair(
-        &audius_reward_manager::id(),
-        &reward_manager.pubkey(),
-        [
-            VERIFY_TRANSFER_SEED_PREFIX.as_bytes().as_ref(),
-            transfer_id.as_ref(),
-        ]
-        .concat()
-        .as_ref(),
-    );
+    let verified_messages_derived_address = get_messages_account(&reward_manager, transfer_id);
 
     let recipient_sol_key = claimable_tokens::utils::program::get_address_pair(
         &claimable_tokens::id(),
@@ -660,80 +449,24 @@ async fn failure_transfer_invalid_message_format() {
 }
 
 #[tokio::test]
+/// Purposefully send a corrupted AAO Attestation
 async fn failure_transfer_invalid_oracle_message_format() {
-    /* Create verified messages and initialize reward manager */
-    let mut program_test = program_test();
 
-    program_test.add_program("claimable_tokens", claimable_tokens::id(), None);
-    let mut rng = thread_rng();
-
-    let mut context = program_test.start_with_context().await;
-
-    let mint = Keypair::new();
-    let mint_authority = Keypair::new();
-
-    let token_account = Keypair::new();
-    let reward_manager = Keypair::new();
-    let manager_account = Keypair::new();
-
-    let rent = context.banks_client.get_rent().await.unwrap();
-
-    create_mint(
-        &mut context,
-        &mint,
-        rent.minimum_balance(spl_token::state::Mint::LEN),
-        &mint_authority.pubkey(),
-    )
-    .await
-    .unwrap();
-
-    init_reward_manager(
-        &mut context,
-        &reward_manager,
-        &token_account,
-        &mint.pubkey(),
-        &manager_account.pubkey(),
-        3,
-    )
-    .await;
-
-    // Generate data and create oracle
-    let key: [u8; 32] = rng.gen();
-    let oracle_priv_key = SecretKey::parse(&key).unwrap();
-    let secp_oracle_pubkey = PublicKey::from_secret_key(&oracle_priv_key);
-    let eth_oracle_address = construct_eth_pubkey(&secp_oracle_pubkey);
-    let oracle_operator: EthereumAddress = rng.gen();
-
-    let (_, oracle_derived_address, _) = find_derived_pair(
-        &audius_reward_manager::id(),
-        &reward_manager.pubkey(),
-        [SENDER_SEED_PREFIX.as_ref(), eth_oracle_address.as_ref()]
-            .concat()
-            .as_ref(),
-    );
-
-    create_sender(
-        &mut context,
-        &reward_manager.pubkey(),
-        &manager_account,
-        eth_oracle_address,
-        oracle_operator,
-    )
-    .await;
-
-    let tokens_amount = 10_000u64;
-    let recipient_eth_key = [7u8; 20];
-    let transfer_id = "4r4t23df32543f55";
-
-    mint_tokens_to(
-        &mut context,
-        &mint.pubkey(),
-        &token_account.pubkey(),
-        &mint_authority,
+    let TestConstants {
+        reward_manager,
+        oracle_priv_key,
+        senders_message,
+        mut context,
+        transfer_id,
+        oracle_derived_address,
+        recipient_eth_key,
+        token_account,
+        mut rng,
+        manager_account,
         tokens_amount,
-    )
-    .await
-    .unwrap();
+        recipient_sol_key,
+        ..
+    } = setup_test_environment().await;
 
     let bot_oracle_message = vote_message!([
         recipient_eth_key.as_ref(),
@@ -744,49 +477,13 @@ async fn failure_transfer_invalid_oracle_message_format() {
     ]
     .concat());
 
-    let senders_message = vote_message!([
-        recipient_eth_key.as_ref(),
-        b"_",
-        tokens_amount.to_le_bytes().as_ref(),
-        b"_",
-        transfer_id.as_ref(),
-        b"_",
-        eth_oracle_address.as_ref(),
-    ]
-    .concat());
-
     // Generate data and create senders
     let keys: [[u8; 32]; 3] = rng.gen();
     let operators: [EthereumAddress; 3] = rng.gen();
     let mut signers: [Pubkey; 3] = unsafe { MaybeUninit::zeroed().assume_init() };
-    for item in keys.iter().enumerate() {
-        let sender_priv_key = SecretKey::parse(item.1).unwrap();
-        let secp_pubkey = PublicKey::from_secret_key(&sender_priv_key);
-        let eth_address = construct_eth_pubkey(&secp_pubkey);
-
-        let (_, derived_address, _) = find_derived_pair(
-            &audius_reward_manager::id(),
-            &reward_manager.pubkey(),
-            [SENDER_SEED_PREFIX.as_ref(), eth_address.as_ref()]
-                .concat()
-                .as_ref(),
-        );
-
-        signers[item.0] = derived_address;
-    }
-
-    for item in keys.iter().enumerate() {
-        let sender_priv_key = SecretKey::parse(item.1).unwrap();
-        let secp_pubkey = PublicKey::from_secret_key(&sender_priv_key);
-        let eth_address = construct_eth_pubkey(&secp_pubkey);
-        create_sender(
-            &mut context,
-            &reward_manager.pubkey(),
-            &manager_account,
-            eth_address,
-            operators[item.0],
-        )
-        .await;
+    for (i, key) in keys.iter().enumerate() {
+        let derived_address = create_sender_from(&reward_manager, &manager_account, &mut context, key, operators[i]).await;
+        signers[i] = derived_address;
     }
 
     let mut instructions = Vec::<Instruction>::new();
@@ -841,26 +538,7 @@ async fn failure_transfer_invalid_oracle_message_format() {
 
     context.banks_client.process_transaction(tx).await.unwrap();
 
-    let (_, verified_messages_derived_address, _) = find_derived_pair(
-        &audius_reward_manager::id(),
-        &reward_manager.pubkey(),
-        [
-            VERIFY_TRANSFER_SEED_PREFIX.as_bytes().as_ref(),
-            transfer_id.as_ref(),
-        ]
-        .concat()
-        .as_ref(),
-    );
-
-    let recipient_sol_key = claimable_tokens::utils::program::get_address_pair(
-        &claimable_tokens::id(),
-        &mint.pubkey(),
-        recipient_eth_key,
-    )
-    .unwrap();
-    println!("Creating...Recipient sol key = {:?}", &recipient_sol_key.derive.address);
-    create_recipient_with_claimable_program(&mut context, &mint.pubkey(), recipient_eth_key).await;
-    println!("Created recipient sol key = {:?}", &recipient_sol_key.derive.address);
+    let verified_messages_derived_address = get_messages_account(&reward_manager, transfer_id);
 
     let tx = Transaction::new_signed_with_payer(
         &[instruction::evaluate_attestations(
@@ -887,132 +565,26 @@ async fn failure_transfer_invalid_oracle_message_format() {
 
 #[tokio::test]
 async fn failure_transfer_incorrect_number_of_verified_messages() {
-    /* Create verified messages and initialize reward manager */
-    let mut program_test = program_test();
-
-    program_test.add_program("claimable_tokens", claimable_tokens::id(), None);
-    let mut rng = thread_rng();
-
-    let mut context = program_test.start_with_context().await;
-
-    let mint = Keypair::new();
-    let mint_authority = Keypair::new();
-
-    let token_account = Keypair::new();
-    let reward_manager = Keypair::new();
-    let manager_account = Keypair::new();
-
-    let rent = context.banks_client.get_rent().await.unwrap();
-
-    create_mint(
-        &mut context,
-        &mint,
-        rent.minimum_balance(spl_token::state::Mint::LEN),
-        &mint_authority.pubkey(),
-    )
-    .await
-    .unwrap();
-
-    init_reward_manager(
-        &mut context,
-        &reward_manager,
-        &token_account,
-        &mint.pubkey(),
-        &manager_account.pubkey(),
-        3,
-    )
-    .await;
-
-    // Generate data and create oracle
-    let key: [u8; 32] = rng.gen();
-    let oracle_priv_key = SecretKey::parse(&key).unwrap();
-    let secp_oracle_pubkey = PublicKey::from_secret_key(&oracle_priv_key);
-    let eth_oracle_address = construct_eth_pubkey(&secp_oracle_pubkey);
-    let oracle_operator: EthereumAddress = rng.gen();
-
-    let (_, oracle_derived_address, _) = find_derived_pair(
-        &audius_reward_manager::id(),
-        &reward_manager.pubkey(),
-        [SENDER_SEED_PREFIX.as_ref(), eth_oracle_address.as_ref()]
-            .concat()
-            .as_ref(),
-    );
-
-    create_sender(
-        &mut context,
-        &reward_manager.pubkey(),
-        &manager_account,
-        eth_oracle_address,
-        oracle_operator,
-    )
-    .await;
-
-    let tokens_amount = 10_000u64;
-    let recipient_eth_key = [7u8; 20];
-    let transfer_id = "4r4t23df32543f55";
-
-    mint_tokens_to(
-        &mut context,
-        &mint.pubkey(),
-        &token_account.pubkey(),
-        &mint_authority,
-        tokens_amount,
-    )
-    .await
-    .unwrap();
-
-    let bot_oracle_message = vote_message!([
-        recipient_eth_key.as_ref(),
-        b"_",
-        tokens_amount.to_le_bytes().as_ref(),
-        b"_",
-        transfer_id.as_ref(),
-    ]
-    .concat());
-
-    let senders_message = vote_message!([
-        recipient_eth_key.as_ref(),
-        b"_",
-        tokens_amount.to_le_bytes().as_ref(),
-        b"_",
-        transfer_id.as_ref(),
-        b"_",
-        eth_oracle_address.as_ref(),
-    ]
-    .concat());
+    let TestConstants {
+        reward_manager,
+        bot_oracle_message,
+        oracle_priv_key,
+        senders_message,
+        mut context,
+        transfer_id,
+        oracle_derived_address,
+        mut rng,
+        manager_account,
+        ..
+     } = setup_test_environment().await;
 
     // Generate data and create senders
     let keys: [[u8; 32]; 3] = rng.gen();
     let operators: [EthereumAddress; 3] = rng.gen();
     let mut signers: [Pubkey; 3] = unsafe { MaybeUninit::zeroed().assume_init() };
-    for item in keys.iter().enumerate() {
-        let sender_priv_key = SecretKey::parse(item.1).unwrap();
-        let secp_pubkey = PublicKey::from_secret_key(&sender_priv_key);
-        let eth_address = construct_eth_pubkey(&secp_pubkey);
-
-        let (_, derived_address, _) = find_derived_pair(
-            &audius_reward_manager::id(),
-            &reward_manager.pubkey(),
-            [SENDER_SEED_PREFIX.as_ref(), eth_address.as_ref()]
-                .concat()
-                .as_ref(),
-        );
-
-        signers[item.0] = derived_address;
-    }
-
-    for item in keys.iter().enumerate() {
-        let sender_priv_key = SecretKey::parse(item.1).unwrap();
-        let secp_pubkey = PublicKey::from_secret_key(&sender_priv_key);
-        let eth_address = construct_eth_pubkey(&secp_pubkey);
-        create_sender(
-            &mut context,
-            &reward_manager.pubkey(),
-            &manager_account,
-            eth_address,
-            operators[item.0],
-        )
-        .await;
+    for (i, key) in keys.iter().enumerate() {
+        let derived_address = create_sender_from(&reward_manager, &manager_account, &mut context, key, operators[i]).await;
+        signers[i] = derived_address;
     }
 
     let mut instructions = Vec::<Instruction>::new();
@@ -1092,17 +664,6 @@ fn get_messages_account(reward_manager: &Keypair, transfer_id: &str) -> Pubkey {
         .as_ref(),
     );
     verified_messages_derived_address
-}
-
-fn get_oracle_address(reward_manager: &Keypair, eth_oracle_address: [u8; 20]) -> Pubkey {
-    let (_, oracle_derived_address, _) = find_derived_pair(
-        &audius_reward_manager::id(),
-        &reward_manager.pubkey(),
-        [SENDER_SEED_PREFIX.as_ref(), eth_oracle_address.as_ref()]
-            .concat()
-            .as_ref(),
-    );
-    oracle_derived_address
 }
 
 async fn create_sender_from(
