+++ solana-programs/reward-manager/program/tests/utils.rs
@@ -1,8 +1,10 @@
 #![cfg(feature = "test-bpf")]
 #![allow(dead_code)]
-use audius_reward_manager::instruction;
-use audius_reward_manager::utils::EthereumAddress;
+
+use audius_reward_manager::{instruction, vote_message, processor::{SENDER_SEED_PREFIX}};
+use audius_reward_manager::utils::{EthereumAddress, find_derived_pair};
 use audius_reward_manager::{id, processor::Processor};
+use claimable_tokens::utils::program::AddressPair;
 use sha3::Digest;
 use solana_program::{
     instruction::Instruction, program_pack::Pack, pubkey::Pubkey, rent::Rent, system_instruction,
@@ -19,6 +21,8 @@ use solana_sdk::{
 };
 use audius_reward_manager::error::AudiusProgramError;
 use solana_program::instruction::InstructionError;
+use rand::{Rng, prelude::ThreadRng, thread_rng};
+use libsecp256k1::{PublicKey, SecretKey};
 
 pub fn program_test() -> ProgramTest {
     ProgramTest::new(
@@ -126,6 +130,18 @@ pub async fn create_sender(
 
     context.banks_client.process_transaction(tx).await.unwrap();
 }
+
+fn get_oracle_address(reward_manager: &Keypair, eth_oracle_address: [u8; 20]) -> Pubkey {
+    let (_, oracle_derived_address, _) = find_derived_pair(
+        &audius_reward_manager::id(),
+        &reward_manager.pubkey(),
+        [SENDER_SEED_PREFIX.as_ref(), eth_oracle_address.as_ref()]
+            .concat()
+            .as_ref(),
+    );
+    oracle_derived_address
+}
+
 
 pub async fn init_reward_manager(
     context: &mut ProgramTestContext,
@@ -309,4 +325,178 @@ pub fn assert_custom_error(res: Result<(), TransportError>, instruction_index: u
         },
         _ => panic!("Expected error")
     }
+}
+
+pub struct TestConstants<'a> {
+    pub reward_manager: Keypair,
+    pub bot_oracle_message: [u8; 128],
+    pub oracle_priv_key: SecretKey,
+    pub senders_message: [u8; 128],
+    pub context: ProgramTestContext,
+    pub transfer_id: &'a str,
+    pub oracle_derived_address: Pubkey,
+    pub mint: Keypair, 
+    pub recipient_eth_key: [u8; 20],
+    pub token_account: Keypair,
+    pub rent: Rent,
+    pub rng: ThreadRng,
+    pub manager_account: Keypair,
+    pub tokens_amount: u64,
+    pub eth_oracle_address: [u8;20], 
+    pub recipient_sol_key: AddressPair,
+    pub min_votes: u8,
+}
+
+pub async fn setup_test_environment<'a>() -> TestConstants<'a> {
+    /* Create verified messages and initialize reward manager */
+    let mut program_test = program_test();
+
+    program_test.add_program("claimable_tokens", claimable_tokens::id(), None);
+    let mut rng = thread_rng();
+
+    let mut context = program_test.start_with_context().await;
+
+    let mint = Keypair::new();
+    let mint_authority = Keypair::new();
+
+    let token_account = Keypair::new();
+    let reward_manager = Keypair::new();
+    let manager_account = Keypair::new();
+    let min_votes: u8 = 3;
+
+    let rent = context.banks_client.get_rent().await.unwrap();
+
+    create_mint(
+        &mut context,
+        &mint,
+        rent.minimum_balance(spl_token::state::Mint::LEN),
+        &mint_authority.pubkey(),
+    )
+    .await
+    .unwrap();
+
+    init_reward_manager(
+        &mut context,
+        &reward_manager,
+        &token_account,
+        &mint.pubkey(),
+        &manager_account.pubkey(),
+        min_votes,
+    )
+    .await;
+
+    // Generate data and create oracle
+    let key: [u8; 32] = rng.gen();
+    let oracle_priv_key = SecretKey::parse(&key).unwrap();
+    let secp_oracle_pubkey = PublicKey::from_secret_key(&oracle_priv_key);
+    let eth_oracle_address = construct_eth_pubkey(&secp_oracle_pubkey);
+    let oracle_operator: EthereumAddress = rng.gen();
+
+    let oracle_derived_address = get_oracle_address(&reward_manager, eth_oracle_address);
+    
+    create_sender(
+        &mut context,
+        &reward_manager.pubkey(),
+        &manager_account,
+        eth_oracle_address,
+        oracle_operator,
+    )
+    .await;
+
+    let tokens_amount = 10_000u64;
+    let recipient_eth_key = [7u8; 20];
+    let transfer_id = "4r4t23df32543f55";
+
+    mint_tokens_to(
+        &mut context,
+        &mint.pubkey(),
+        &token_account.pubkey(),
+        &mint_authority,
+        tokens_amount,
+    )
+    .await
+    .unwrap();
+
+    let bot_oracle_message = vote_message!([
+        recipient_eth_key.as_ref(),
+        b"_",
+        tokens_amount.to_le_bytes().as_ref(),
+        b"_",
+        transfer_id.as_ref(),
+    ]
+    .concat());
+
+    let senders_message = vote_message!([
+        recipient_eth_key.as_ref(),
+        b"_",
+        tokens_amount.to_le_bytes().as_ref(),
+        b"_",
+        transfer_id.as_ref(),
+        b"_",
+        eth_oracle_address.as_ref(),
+    ]
+    .concat());
+
+
+    // Make recipient token account
+
+    let recipient_sol_key = claimable_tokens::utils::program::get_address_pair(
+        &claimable_tokens::id(),
+        &mint.pubkey(),
+        recipient_eth_key,
+    )
+    .unwrap();
+    println!("Creating...Recipient sol key = {:?}", &recipient_sol_key.derive.address);
+    create_recipient_with_claimable_program(&mut context, &mint.pubkey(), recipient_eth_key).await;
+    println!("Created recipient sol key = {:?}", &recipient_sol_key.derive.address);
+
+    return TestConstants {
+        reward_manager,
+        bot_oracle_message,
+        oracle_priv_key,
+        senders_message,
+        context,
+        transfer_id,
+        oracle_derived_address,
+        mint, 
+        recipient_eth_key,
+        token_account,
+        rent,
+        rng,
+        manager_account,
+        tokens_amount,
+        eth_oracle_address, 
+        recipient_sol_key,
+        min_votes,
+    }
+}
+
+pub async fn create_sender_from(
+    reward_manager: &Keypair,
+    manager_account: &Keypair,
+    context: &mut ProgramTestContext,
+    key: &[u8; 32],
+    operator: [u8; 20]) -> Pubkey {
+    let sender_priv_key = SecretKey::parse(key).unwrap();
+    let secp_pubkey = PublicKey::from_secret_key(&sender_priv_key);
+    let eth_address = construct_eth_pubkey(&secp_pubkey);
+
+    let (_, derived_address, _) = find_derived_pair(
+        &audius_reward_manager::id(),
+        &reward_manager.pubkey(),
+        [SENDER_SEED_PREFIX.as_ref(), eth_address.as_ref()]
+            .concat()
+            .as_ref(),
+    );
+
+    create_sender(
+        context,
+        &reward_manager.pubkey(),
+        &manager_account,
+        eth_address,
+        operator
+    )
+    .await;
+
+    derived_address
 }
